# 8299061: Using lambda to optimize GraphKit::compute_stack_effects()